# Piece picking perf improvements