--- a/src/MonoTorrent.PiecePicking/MonoTorrent.PiecePicking/StandardPicker.cs
+++ b/src/MonoTorrent.PiecePicking/MonoTorrent.PiecePicking/StandardPicker.cs
@@ -43,7 +43,8 @@ namespace MonoTorrent.PiecePicking
 
         readonly Dictionary<int, List<Piece>> duplicates;
         readonly Dictionary<int, Piece> requests;
-        MutableBitField alreadyRequested;
+        MutableBitField alreadyRequestedBitField;
+        MutableBitField canRequestBitField;
 
         ITorrentData TorrentData { get; set; }
 
@@ -107,7 +108,7 @@ namespace MonoTorrent.PiecePicking
             }
 
             while (toRemove.Count > 0) {
-                alreadyRequested[toRemove.Peek ()] = false;
+                alreadyRequestedBitField[toRemove.Peek ()] = false;
                 requests.Remove (toRemove.Pop ());
             }
             return count;
@@ -144,7 +145,8 @@ namespace MonoTorrent.PiecePicking
         public void Initialise (ITorrentData torrentData)
         {
             TorrentData = torrentData;
-            alreadyRequested = new MutableBitField (torrentData.PieceCount ());
+            alreadyRequestedBitField = new MutableBitField (torrentData.PieceCount ());
+            canRequestBitField = new MutableBitField (torrentData.PieceCount ());
             requests.Clear ();
         }
 
@@ -193,6 +195,8 @@ namespace MonoTorrent.PiecePicking
 
         public void Reset ()
         {
+            alreadyRequestedBitField.SetAll (false);
+            canRequestBitField.SetAll (false);
             duplicates.Clear ();
             requests.Clear ();
         }
@@ -216,7 +220,7 @@ namespace MonoTorrent.PiecePicking
             // ensure our received piece is not re-requested again.
             if (!duplicates.TryGetValue (request.PieceIndex, out List<Piece> extraPieces)) {
                 if (pieceComplete) {
-                    alreadyRequested[request.PieceIndex] = false;
+                    alreadyRequestedBitField[request.PieceIndex] = false;
                     requests.Remove (request.PieceIndex);
                     PieceCache.Enqueue (primaryPiece);
                 }
@@ -241,7 +245,7 @@ namespace MonoTorrent.PiecePicking
 
             // If the piece is complete then remove it, and any dupes, from the picker.
             if (pieceComplete) {
-                alreadyRequested[request.PieceIndex] = false;
+                alreadyRequestedBitField[request.PieceIndex] = false;
                 requests.Remove (request.PieceIndex);
                 duplicates.Remove (primaryPiece.Index);
 
@@ -359,14 +363,13 @@ namespace MonoTorrent.PiecePicking
 
             for (int i = 0; i < pieces.Count; i++) {
                 int index = pieces[i];
-                // A peer should only suggest a piece he has, but just in case.
-                if (index >= bitfield.Length || !bitfield[index] || alreadyRequested [index])
+                if (index >= bitfield.Length || !bitfield[index] || alreadyRequestedBitField [index])
                     continue;
 
                 pieces.RemoveAt (i);
                 var p = PieceCache.Dequeue ().Initialise (index, TorrentData.BytesPerPiece (index));
                 requests.Add (p.Index, p);
-                alreadyRequested[p.Index] = true;
+                alreadyRequestedBitField[p.Index] = true;
                 return p.Blocks[0].CreateRequest (peer);
             }
 
@@ -390,7 +393,7 @@ namespace MonoTorrent.PiecePicking
                 // Request the piece
                 var p = PieceCache.Dequeue ().Initialise (checkIndex + i, TorrentData.BytesPerPiece (checkIndex + i));
                 this.requests.Add (p.Index, p);
-                alreadyRequested[p.Index] = true;
+                alreadyRequestedBitField[p.Index] = true;
                 for (int j = 0; j < p.Blocks.Length && totalRequested < requests.Length; j++)
                     requests[totalRequested++] = p.Blocks[j].CreateRequest (peer);
             }
@@ -401,19 +404,7 @@ namespace MonoTorrent.PiecePicking
         {
             // This is the easiest case to consider - special case it
             if (pieceCount == 1) {
-                while (pieceStartIndex <= pieceEndIndex && (pieceStartIndex = bitfield.FirstTrue (pieceStartIndex, pieceEndIndex)) != -1) {
-                    var end = bitfield.FirstFalse (pieceStartIndex, pieceEndIndex);
-
-                    // If end is a valid value, it's the first *false* piece. Subtract '1' from it
-                    // to give us the last available piece we can request. If it's -1 then we can use
-                    // 'pieceEndIndex' as the last available piece to request as all pieces are available.
-                    var lastAvailable = end == -1 ? pieceEndIndex : end - 1;
-                    for (int i = pieceStartIndex; i <= lastAvailable; i++)
-                        if (!alreadyRequested[i])
-                            return i;
-                    pieceStartIndex = lastAvailable + 1;
-                }
-                return -1;
+                return canRequestBitField.From (bitfield).NAnd (alreadyRequestedBitField).FirstTrue (pieceStartIndex, pieceEndIndex);
             }
 
             int largestStart = 0;
@@ -425,7 +416,7 @@ namespace MonoTorrent.PiecePicking
 
                 // Do not include 'end' as it's the first *false* piece.
                 for (int i = pieceStartIndex; i < end; i++)
-                    if (alreadyRequested [i])
+                    if (alreadyRequestedBitField [i])
                         end = i;
 
                 if ((end - pieceStartIndex) >= pieceCount)