The delta between the first commit where the benchmark was added and the final performance improving commit is (currently):

Unoptimised code:
|                                     Method |     Mean |     Error |    StdDev | Allocated |
|------------------------------------------- |---------:|----------:|----------:|----------:|
|                            PickAndValidate | 8.229 ms | 0.0331 ms | 0.0310 ms |     16 KB |
|              PickAndValidate_600Concurrent | 7.451 ms | 0.0507 ms | 0.0475 ms |     16 KB |
| PickAndValidate_600Concurrent_60Requesters | 7.361 ms | 0.0313 ms | 0.0293 ms |     16 KB |


Optimised code:
|                                     Method |     Mean |   Error |  StdDev |  Gen 0 | Allocated |
|------------------------------------------- |---------:|--------:|--------:|-------:|----------:|
|                            PickAndValidate | 689.1 us | 2.91 us | 2.72 us | 2.9297 |     31 KB |
|              PickAndValidate_600Concurrent | 633.3 us | 2.50 us | 2.34 us | 2.9297 |     32 KB |
| PickAndValidate_600Concurrent_60Requesters | 672.2 us | 3.49 us | 2.91 us | 2.9297 |     32 KB |